Move WidgetFrame back for now
