+++ static/app/views/projectDetail/projectScoreCards/actionWrapper.tsx
@@ -3,5 +3,5 @@ import styled from '@emotion/styled';
 import {space} from 'sentry/styles/space';
 
 export const ActionWrapper = styled('div')`
-  padding: ${space(1)} 0 0 0;
+  padding: ${space(2)};
 `;
