+++ static/app/views/projectDetail/projectScoreCards/projectStabilityScoreCard.tsx
@@ -13,7 +13,7 @@ import {getPeriod} from 'sentry/utils/duration/getPeriod';
 import {useApiQuery} from 'sentry/utils/queryClient';
 import useOrganization from 'sentry/utils/useOrganization';
 import {BigNumberWidget} from 'sentry/views/dashboards/widgets/bigNumberWidget/bigNumberWidget';
-import {WidgetFrame} from 'sentry/views/dashboards/widgets/widgetFrame/widgetFrame';
+import {WidgetFrame} from 'sentry/views/dashboards/widgets/common/widgetFrame';
 import {
   getSessionTermDescription,
   SessionTerm,
